Add missing block header and body fields
The block header is updated to allow for the P2P spec to be compatible
with the RPC spec. Before these changes, p2p spec was not enough to
serve to block RPC requests.

Block bodies include Transactions and Receipts because RPC requires the
block to have transactions and receipts. A peer could retrieve them in a
separate request, however, since block bodies are retrieved over
multiple messages I don't see much benefit in requesting them
separately.

Also, note Events were added to Receipt common because `BlockID` is not
enough to determine which event belongs to which receipt. We require the
events for the bloom filter.