Add `deployed_contracts` and `replaced_classes` to `StateDiff`
To verify the state diff commitment we need to have access to all the
fields present in the `StateDiff` returned by the feeder gateway.
Therefore, we need to have `deployed_contract` and `replaced_classes` in
the `StateDiff` messages.

The reason for not including the `DeclaredV0Classes` and
`DeclaredV1Classes` is because these can be derived from the `Classes`
message, which all Cairo0 and Cairo1 classes can be mapped to
`DeclaredV0Classes` and `DeclaredV1Classes`, respectively.