L1 handler message hash values are 256 bit
This change introduces a new `Hash256` message type to describe fields
that are hash values but cannot be represented as a Felt252.

For example, message hashes for L1 messages are defined as Keccak256 hashes:
https://docs.starknet.io/architecture-and-concepts/network-architecture/messaging-mechanism/#hashing_l1-l2

All other uses of the `Hash` message are _indeed_ hash function output either
from the Starknet Poseidon or Pedersen hashes, so `Hash` should be parsed
as a Felt252 value.